Move download_file function to bot_utils.py
